fix(diagnostics): sépare les types d'inspection du service serveur
Le composant client tirait le service, donc le client Postgres : la page ne compilait plus.
