+++ src/features/parcours/dossiers-ds/services/inspection.service.ts
@@ -3,6 +3,7 @@ import { db } from "@/shared/database/client";
 import { parcoursPrevention, users } from "@/shared/database/schema";
 import { graphqlClient } from "../adapters/graphql/client";
 import type { DossierInspection } from "../adapters/graphql/types";
+import type { CandidatDemandeur, InspectionDossier, MotifRapprochement } from "../domain/types/inspection.types";
 
 /**
  * Aide à identifier le demandeur d'un dossier DN qu'aucun parcours ne suit (ADR-0027).
@@ -11,43 +12,6 @@ import type { DossierInspection } from "../adapters/graphql/types";
  * une adresse e-mail peut être celle de l'AMO, un nom peut être porté par deux personnes. La
  * décision reste humaine, l'écran ne fait que lui épargner l'enquête.
  */
-
-/** Ce qui a fait correspondre un demandeur, du plus fiable au moins fiable. */
-export type MotifRapprochement = "adresse_logement" | "telephone" | "email" | "nom_prenom";
-
-export const MOTIF_LABELS: Record<MotifRapprochement, string> = {
-  adresse_logement: "même adresse de logement",
-  telephone: "même téléphone",
-  email: "même adresse e-mail",
-  nom_prenom: "mêmes nom et prénom",
-};
-
-export interface CandidatDemandeur {
-  parcoursId: string;
-  userId: string;
-  nom: string | null;
-  prenom: string | null;
-  email: string | null;
-  currentStep: string;
-  adresse: string | null;
-  motifs: MotifRapprochement[];
-}
-
-export interface InspectionDossier {
-  dn: {
-    number: number;
-    state: string;
-    dateDepot?: string;
-    emailUsager: string | null;
-    nomDeclare: string | null;
-    prenomDeclare: string | null;
-    deposeParUnTiers: boolean;
-    mandataire: string | null;
-    /** Champs du formulaire non vides, tels que DN les renvoie. */
-    champs: Array<{ label: string; valeur: string }>;
-  };
-  candidats: CandidatDemandeur[];
-}
 
 /** Derniers chiffres d'un numéro, pour comparer des formats de saisie hétérogènes. */
 function chiffresSignificatifs(telephone: string | null | undefined): string | null {
